ROC-7376 Extract staff email from support
