+++ src/main/java/uk/gov/hmcts/cmc/claimstore/controllers/support/SupportController.java
@@ -21,29 +21,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ServerErrorException;
-import uk.gov.hmcts.cmc.claimstore.events.ccj.CCJStaffNotificationHandler;
-import uk.gov.hmcts.cmc.claimstore.events.ccj.CountyCourtJudgmentEvent;
 import uk.gov.hmcts.cmc.claimstore.events.claim.CitizenClaimCreatedEvent;
-import uk.gov.hmcts.cmc.claimstore.events.claim.CitizenClaimIssuedEvent;
-import uk.gov.hmcts.cmc.claimstore.events.claim.DocumentGenerator;
 import uk.gov.hmcts.cmc.claimstore.events.claim.PostClaimOrchestrationHandler;
-import uk.gov.hmcts.cmc.claimstore.events.claimantresponse.ClaimantResponseEvent;
-import uk.gov.hmcts.cmc.claimstore.events.claimantresponse.ClaimantResponseStaffNotificationHandler;
-import uk.gov.hmcts.cmc.claimstore.events.offer.AgreementCountersignedEvent;
-import uk.gov.hmcts.cmc.claimstore.events.offer.AgreementCountersignedStaffNotificationHandler;
-import uk.gov.hmcts.cmc.claimstore.events.paidinfull.PaidInFullEvent;
-import uk.gov.hmcts.cmc.claimstore.events.paidinfull.PaidInFullStaffNotificationHandler;
-import uk.gov.hmcts.cmc.claimstore.events.response.DefendantResponseEvent;
-import uk.gov.hmcts.cmc.claimstore.events.response.DefendantResponseStaffNotificationHandler;
-import uk.gov.hmcts.cmc.claimstore.events.response.MoreTimeRequestedEvent;
-import uk.gov.hmcts.cmc.claimstore.events.response.MoreTimeRequestedStaffNotificationHandler;
 import uk.gov.hmcts.cmc.claimstore.events.solicitor.RepresentedClaimCreatedEvent;
-import uk.gov.hmcts.cmc.claimstore.events.solicitor.RepresentedClaimIssuedEvent;
-import uk.gov.hmcts.cmc.claimstore.exceptions.ConflictException;
 import uk.gov.hmcts.cmc.claimstore.exceptions.NotFoundException;
-import uk.gov.hmcts.cmc.claimstore.idam.models.GeneratePinResponse;
 import uk.gov.hmcts.cmc.claimstore.idam.models.User;
-import uk.gov.hmcts.cmc.claimstore.idam.models.UserDetails;
 import uk.gov.hmcts.cmc.claimstore.rules.ClaimSubmissionOperationIndicatorRule;
 import uk.gov.hmcts.cmc.claimstore.services.ClaimService;
 import uk.gov.hmcts.cmc.claimstore.services.IntentionToProceedService;
@@ -55,24 +37,14 @@ import uk.gov.hmcts.cmc.domain.models.Claim;
 import uk.gov.hmcts.cmc.domain.models.ClaimDocumentType;
 import uk.gov.hmcts.cmc.domain.models.ClaimSubmissionOperationIndicators;
 import uk.gov.hmcts.cmc.domain.models.MediationRequest;
-import uk.gov.hmcts.cmc.domain.models.claimantresponse.ClaimantResponse;
-import uk.gov.hmcts.cmc.domain.models.claimantresponse.FormaliseOption;
-import uk.gov.hmcts.cmc.domain.models.claimantresponse.ResponseAcceptation;
-import uk.gov.hmcts.cmc.domain.models.response.Response;
 import uk.gov.hmcts.cmc.domain.utils.LocalDateTimeFactory;
-import uk.gov.hmcts.cmc.domain.utils.ResponseUtils;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.function.Predicate;
 import java.util.function.Supplier;
 
 import static java.lang.String.format;
-import static uk.gov.hmcts.cmc.claimstore.utils.CommonErrors.MISSING_CLAIMANT_RESPONSE;
 import static uk.gov.hmcts.cmc.claimstore.utils.CommonErrors.MISSING_REPRESENTATIVE;
-import static uk.gov.hmcts.cmc.claimstore.utils.CommonErrors.MISSING_RESPONSE;
-import static uk.gov.hmcts.cmc.domain.models.claimantresponse.ClaimantResponseType.ACCEPTATION;
-import static uk.gov.hmcts.cmc.domain.models.claimantresponse.ClaimantResponseType.REJECTION;
 
 @RestController
 @RequestMapping("/support")
@@ -80,19 +52,12 @@ public class SupportController {
 
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
-    private static final String CLAIM = "Claim ";
+    public static final String CLAIM = "Claim ";
     public static final String CLAIM_DOES_NOT_EXIST = "Claim %s does not exist";
-    private static final String AUTHORISATION_IS_REQUIRED = "Authorisation is required";
+    public static final String AUTHORISATION_IS_REQUIRED = "Authorisation is required";
 
     private final ClaimService claimService;
     private final UserService userService;
-    private final DocumentGenerator documentGenerator;
-    private final MoreTimeRequestedStaffNotificationHandler moreTimeRequestedStaffNotificationHandler;
-    private final DefendantResponseStaffNotificationHandler defendantResponseStaffNotificationHandler;
-    private final CCJStaffNotificationHandler ccjStaffNotificationHandler;
-    private final AgreementCountersignedStaffNotificationHandler agreementCountersignedStaffNotificationHandler;
-    private final ClaimantResponseStaffNotificationHandler claimantResponseStaffNotificationHandler;
-    private final PaidInFullStaffNotificationHandler paidInFullStaffNotificationHandler;
     private final DocumentsService documentsService;
     private final PostClaimOrchestrationHandler postClaimOrchestrationHandler;
     private final MediationReportService mediationReportService;
@@ -101,30 +66,16 @@ public class SupportController {
 
     @SuppressWarnings("squid:S00107")
     public SupportController(
-            ClaimService claimService,
-            UserService userService,
-            DocumentGenerator documentGenerator,
-            MoreTimeRequestedStaffNotificationHandler moreTimeRequestedStaffNotificationHandler,
-            DefendantResponseStaffNotificationHandler defendantResponseStaffNotificationHandler,
-            CCJStaffNotificationHandler ccjStaffNotificationHandler,
-            AgreementCountersignedStaffNotificationHandler agreementCountersignedStaffNotificationHandler,
-            ClaimantResponseStaffNotificationHandler claimantResponseStaffNotificationHandler,
-            PaidInFullStaffNotificationHandler paidInFullStaffNotificationHandler,
-            DocumentsService documentsService,
-            PostClaimOrchestrationHandler postClaimOrchestrationHandler,
-            MediationReportService mediationReportService,
-            ClaimSubmissionOperationIndicatorRule claimSubmissionOperationIndicatorRule,
-            IntentionToProceedService intentionToProceedService
+        ClaimService claimService,
+        UserService userService,
+        DocumentsService documentsService,
+        PostClaimOrchestrationHandler postClaimOrchestrationHandler,
+        MediationReportService mediationReportService,
+        ClaimSubmissionOperationIndicatorRule claimSubmissionOperationIndicatorRule,
+        IntentionToProceedService intentionToProceedService
     ) {
         this.claimService = claimService;
         this.userService = userService;
-        this.documentGenerator = documentGenerator;
-        this.moreTimeRequestedStaffNotificationHandler = moreTimeRequestedStaffNotificationHandler;
-        this.defendantResponseStaffNotificationHandler = defendantResponseStaffNotificationHandler;
-        this.ccjStaffNotificationHandler = ccjStaffNotificationHandler;
-        this.agreementCountersignedStaffNotificationHandler = agreementCountersignedStaffNotificationHandler;
-        this.claimantResponseStaffNotificationHandler = claimantResponseStaffNotificationHandler;
-        this.paidInFullStaffNotificationHandler = paidInFullStaffNotificationHandler;
         this.documentsService = documentsService;
         this.postClaimOrchestrationHandler = postClaimOrchestrationHandler;
         this.mediationReportService = mediationReportService;
@@ -132,60 +83,18 @@ public class SupportController {
         this.intentionToProceedService = intentionToProceedService;
     }
 
-    @PutMapping("/claim/{referenceNumber}/event/{event}/resend-staff-notifications")
-    @ApiOperation("Resend staff notifications associated with provided event")
-    public void resendStaffNotifications(
-            @PathVariable("referenceNumber") String referenceNumber,
-            @PathVariable("event") String event
-    ) {
-        User user = userService.authenticateAnonymousCaseWorker();
-        String authorisation = user.getAuthorisation();
-
-        Claim claim = claimService.getClaimByReferenceAnonymous(referenceNumber)
-            .orElseThrow(claimNotFoundException(referenceNumber));
-
-        switch (event) {
-            case "claim":
-                resendStaffNotificationsOnClaimIssued(claim, authorisation);
-                break;
-            case "more-time":
-                resendStaffNotificationOnMoreTimeRequested(claim);
-                break;
-            case "response":
-                resendStaffNotificationOnDefendantResponseSubmitted(claim, authorisation);
-                break;
-            case "ccj":
-                resendStaffNotificationCCJRequestSubmitted(claim, authorisation);
-                break;
-            case "settlement":
-                resendStaffNotificationOnAgreementCountersigned(claim, authorisation);
-                break;
-            case "claimant-response":
-                resendStaffNotificationClaimantResponse(claim, authorisation);
-                break;
-            case "intent-to-proceed":
-                resendStaffNotificationForIntentToProceed(claim, authorisation);
-                break;
-            case "paid-in-full":
-                resendStaffNotificationForPaidInFull(claim);
-                break;
-            default:
-                throw new NotFoundException("Event " + event + " is not supported");
-        }
-    }
-
     @PutMapping("/documents/{referenceNumber}/{documentType}")
     @ApiOperation("Ensure a document is available on CCD")
     @ApiResponses({
-            @ApiResponse(code = 200, message = "OK"),
-            @ApiResponse(code = 201, message = "Created"),
-            @ApiResponse(code = 404, message = "Claim not found"),
-            @ApiResponse(code = 500, message = "Unable to upload document")
+        @ApiResponse(code = 200, message = "OK"),
+        @ApiResponse(code = 201, message = "Created"),
+        @ApiResponse(code = 404, message = "Claim not found"),
+        @ApiResponse(code = 500, message = "Unable to upload document")
     })
     @SuppressWarnings("squid:S2201") // orElseThrow does not ignore the result
     public ResponseEntity<?> uploadDocumentToDocumentManagement(
-            @PathVariable("referenceNumber") String referenceNumber,
-            @PathVariable("documentType") ClaimDocumentType documentType
+        @PathVariable("referenceNumber") String referenceNumber,
+        @PathVariable("documentType") ClaimDocumentType documentType
     ) {
         User caseworker = userService.authenticateAnonymousCaseWorker();
 
@@ -200,12 +109,12 @@ public class SupportController {
 
         // local claim object is now outdated
         claimService.getClaimByReferenceAnonymous(referenceNumber)
-                .orElseThrow(() -> new IllegalStateException("Missing claim " + referenceNumber))
-                .getClaimDocument(documentType)
-                .orElseThrow(() -> new ServerErrorException(
-                        "Unable to upload the document. Please try again later",
-                        new NotFoundException("Unable to upload the document. Please try again later")
-                ));
+            .orElseThrow(() -> new IllegalStateException("Missing claim " + referenceNumber))
+            .getClaimDocument(documentType)
+            .orElseThrow(() -> new ServerErrorException(
+                "Unable to upload the document. Please try again later",
+                new NotFoundException("Unable to upload the document. Please try again later")
+            ));
 
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
@@ -213,9 +122,9 @@ public class SupportController {
     @PutMapping("/claim/{referenceNumber}/reset-operation")
     @ApiOperation("Redo any failed operation. Use the claim submission indicators to indicate the operation to redo.")
     public void resetOperation(
-            @PathVariable("referenceNumber") String referenceNumber,
-            @RequestBody ClaimSubmissionOperationIndicators indicators,
-            @RequestHeader(value = HttpHeaders.AUTHORIZATION) String authorisation
+        @PathVariable("referenceNumber") String referenceNumber,
+        @RequestBody ClaimSubmissionOperationIndicators indicators,
+        @RequestHeader(value = HttpHeaders.AUTHORIZATION) String authorisation
     ) {
         if (StringUtils.isBlank(authorisation)) {
             throw new BadRequestException(AUTHORISATION_IS_REQUIRED);
@@ -261,8 +170,8 @@ public class SupportController {
     @PostMapping(value = "/sendMediation", consumes = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation("Generate and Send Mediation Report for Telephone Mediation Service")
     public void sendMediation(
-            @RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String authorisation,
-            @RequestBody MediationRequest mediationRequest
+        @RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String authorisation,
+        @RequestBody MediationRequest mediationRequest
     ) {
         mediationReportService.sendMediationReport(authorisation, mediationRequest.getReportDate());
 
@@ -271,10 +180,10 @@ public class SupportController {
     @PostMapping(value = "/claims/checkIntentionToProceedDeadline")
     @ApiOperation("Stay claims past their intention proceed deadline")
     public void checkClaimsPastIntentionToProceedDeadline(
-            @RequestHeader(value = HttpHeaders.AUTHORIZATION) String authorisation,
-            @RequestParam(required = false)
-            @ApiParam("Optional. If supplied check will run as if triggered at this timestamp. Format is ISO 8601")
-            @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime localDateTime
+        @RequestHeader(value = HttpHeaders.AUTHORIZATION) String authorisation,
+        @RequestParam(required = false)
+        @ApiParam("Optional. If supplied check will run as if triggered at this timestamp. Format is ISO 8601")
+        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime localDateTime
     ) {
 
         LocalDateTime runDateTime = localDateTime == null ? LocalDateTimeFactory.nowInLocalZone() : localDateTime;
@@ -282,114 +191,8 @@ public class SupportController {
         User user = userService.getUser(authorisation);
         String format = runDateTime.format(DateTimeFormatter.ofPattern("dd-MM-yyyy hh:mm:ss"));
         logger.info(format("checkClaimsPastIntentionToProceedDeadline called by %s for date: %s",
-                user.getUserDetails().getId(), format));
+            user.getUserDetails().getId(), format));
         intentionToProceedService.checkClaimsPastIntentionToProceedDeadline(runDateTime, user);
-    }
-
-    private void resendStaffNotificationCCJRequestSubmitted(Claim claim, String authorisation) {
-        this.ccjStaffNotificationHandler.onDefaultJudgmentRequestSubmitted(
-                new CountyCourtJudgmentEvent(claim, authorisation)
-        );
-    }
-
-    private void resendStaffNotificationsOnClaimIssued(Claim claim, String authorisation) {
-        if (StringUtils.isBlank(authorisation)) {
-            throw new BadRequestException(AUTHORISATION_IS_REQUIRED);
-        }
-        if (claim.getDefendantId() != null) {
-            throw new ConflictException("Claim has already been linked to defendant - cannot send notification");
-        }
-
-        UserDetails userDetails = userService.getUserDetails(authorisation);
-        String fullName = userDetails.getFullName();
-
-        if (!claim.getClaimData().isClaimantRepresented()) {
-            GeneratePinResponse pinResponse = userService.generatePin(
-                claim.getClaimData().getDefendant().getName(),
-                authorisation
-            );
-
-            claimService.linkLetterHolder(claim, pinResponse.getUserId(), authorisation);
-            documentGenerator.generateForNonRepresentedClaim(
-                    new CitizenClaimIssuedEvent(claim, pinResponse.getPin(), fullName, authorisation)
-            );
-        } else {
-            documentGenerator.generateForRepresentedClaim(
-                    new RepresentedClaimIssuedEvent(claim, fullName, authorisation)
-            );
-        }
-    }
-
-    private void resendStaffNotificationForIntentToProceed(Claim claim, String authorization) {
-        ClaimantResponse claimantResponse = claim.getClaimantResponse()
-            .orElseThrow(() -> new IllegalArgumentException(MISSING_CLAIMANT_RESPONSE));
-
-        if (claimantResponse.getType() != REJECTION) {
-            throw new IllegalArgumentException("Rejected Claimant Response is mandatory for 'intent-to-proceed' event");
-        }
-
-        claimantResponseStaffNotificationHandler.notifyStaffWithClaimantsIntentionToProceed(
-            new ClaimantResponseEvent(claim, authorization)
-        );
-    }
-
-    private void resendStaffNotificationOnMoreTimeRequested(Claim claim) {
-        if (!claim.isMoreTimeRequested()) {
-            throw new ConflictException("More time has not been requested yet - cannot send notification");
-        }
-
-        // Defendant email is not available at this point however it is not used in staff notifications
-        MoreTimeRequestedEvent event = new MoreTimeRequestedEvent(claim, claim.getResponseDeadline(), null);
-        moreTimeRequestedStaffNotificationHandler.sendNotifications(event);
-    }
-
-    private void resendStaffNotificationOnDefendantResponseSubmitted(Claim claim, String authorization) {
-        if (!claim.getResponse().isPresent()) {
-            throw new ConflictException(CLAIM + claim.getReferenceNumber() + " does not have associated response");
-        }
-        DefendantResponseEvent event = new DefendantResponseEvent(claim, authorization);
-        defendantResponseStaffNotificationHandler.onDefendantResponseSubmitted(event);
-    }
-
-    private void resendStaffNotificationOnAgreementCountersigned(Claim claim, String authorisation) {
-        if (claim.getSettlementReachedAt() == null) {
-            throw new ConflictException(CLAIM + claim.getReferenceNumber() + " does not have a settlement");
-        }
-        AgreementCountersignedEvent event = new AgreementCountersignedEvent(claim, null, authorisation);
-        agreementCountersignedStaffNotificationHandler.onAgreementCountersigned(event);
-    }
-
-    private void resendStaffNotificationClaimantResponse(Claim claim, String authorization) {
-        ClaimantResponse claimantResponse = claim.getClaimantResponse()
-                .orElseThrow(() -> new IllegalArgumentException(MISSING_CLAIMANT_RESPONSE));
-        if (!isSettlementAgreement(claim, claimantResponse)) {
-            claimantResponseStaffNotificationHandler.onClaimantResponse(
-                new ClaimantResponseEvent(claim, authorization)
-            );
-        }
-    }
-
-    @SuppressWarnings("squid:S2201") // not ignored
-    private void resendStaffNotificationForPaidInFull(Claim claim) {
-        claim.getMoneyReceivedOn()
-            .orElseThrow(() -> new IllegalArgumentException("Claim missing money received on date"));
-        paidInFullStaffNotificationHandler.onPaidInFullEvent(new PaidInFullEvent(claim));
-    }
-
-    private boolean isSettlementAgreement(Claim claim, ClaimantResponse claimantResponse) {
-        Response response = claim.getResponse().orElseThrow(() -> new IllegalStateException(MISSING_RESPONSE));
-
-        if (shouldFormaliseResponseAcceptance(response, claimantResponse)) {
-            return ((ResponseAcceptation) claimantResponse).getFormaliseOption()
-                    .filter(Predicate.isEqual(FormaliseOption.SETTLEMENT)).isPresent();
-        }
-        return false;
-    }
-
-    private boolean shouldFormaliseResponseAcceptance(Response response, ClaimantResponse claimantResponse) {
-        return ACCEPTATION == claimantResponse.getType()
-                && !ResponseUtils.isResponseStatesPaid(response)
-                && !ResponseUtils.isResponsePartAdmitPayImmediately(response);
     }
 
     private Supplier<NotFoundException> claimNotFoundException(String reference) {
